Let's rebel with more than one class in each file
