+++ Filesystem.Akka.Tests/Filesystem.Akka.Tests.csproj
@@ -65,9 +65,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="CreateFolder.Tests.cs" />
-    <Compile Include="DeleteFileExists.Tests.cs" />
-    <Compile Include="DeleteFileMissing.Tests.cs" />
-    <Compile Include="DeleteFolderPopulated.Tests.cs" />
+    <Compile Include="DeleteFile.Tests.cs" />
     <Compile Include="EmptyFolder.Tests.cs" />
     <Compile Include="FileExists.Tests.cs" />
     <Compile Include="FolderExists.Tests.cs" />
@@ -79,7 +77,7 @@
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
-    <Compile Include="DeleteFolderEmpty.Tests.cs" />
+    <Compile Include="DeleteFolder.Tests.cs" />
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
